estimator/forecast_service/pyproject.toml
[tool.poetry]
name = "forecast_service"
version = "0.1.0"
description = ""
authors = ["Alessandro Pacielli <alessandropacielli@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
tensorflow = "^2.1.0"
pandas = "numpy"
numpy = "^1.18.2"
influxdb = "^5.3.0"
keras = "^2.3.1"
kafka-python = "^2.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.4.1"
autopep8 = "^1.5.1"
mypy = "^0.770"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
